refactor: [#260] move get_next_step_guidance() to view layer
