+++ src/presentation/views/commands/show/environment_info.rs
@@ -31,7 +31,7 @@ use crate::application::command_handlers::show::info::EnvironmentInfo;
 ///     "Created".to_string(),
 ///     "LXD".to_string(),
 ///     created_at,
-///     "Next: Run 'provision my-env' to create infrastructure".to_string(),
+///     "created".to_string(),
 /// );
 ///
 /// let output = EnvironmentInfoView::render(&info);
@@ -73,7 +73,7 @@ impl EnvironmentInfoView {
     ///     "Provisioned".to_string(),
     ///     "LXD".to_string(),
     ///     Utc::now(),
-    ///     "Next: Run 'configure prod-env' to install software".to_string(),
+    ///     "provisioned".to_string(),
     /// ).with_infrastructure(InfrastructureInfo::new(
     ///     IpAddr::V4(Ipv4Addr::new(10, 140, 190, 171)),
     ///     22,
@@ -140,9 +140,45 @@ impl EnvironmentInfoView {
 
         // Next step guidance
         lines.push(String::new()); // blank line
-        lines.push(info.next_step.clone());
+        lines.push(Self::get_next_step_guidance(&info.state_name));
 
         lines.join("\n")
+    }
+
+    /// Get next step guidance based on current state
+    fn get_next_step_guidance(state_name: &str) -> String {
+        match state_name {
+            "created" => "Run 'provision' to create infrastructure.".to_string(),
+            "provisioning" => {
+                "Provisioning in progress. Wait for completion or check logs.".to_string()
+            }
+            "provisioned" => "Run 'configure' to set up the system.".to_string(),
+            "configuring" => {
+                "Configuration in progress. Wait for completion or check logs.".to_string()
+            }
+            "configured" => "Run 'release' to deploy the tracker software.".to_string(),
+            "releasing" => "Release in progress. Wait for completion or check logs.".to_string(),
+            "released" => "Run 'run' to start the tracker services.".to_string(),
+            "running" => "Services are running. Use 'test' to verify health.".to_string(),
+            "destroying" => "Destruction in progress. Wait for completion.".to_string(),
+            "destroyed" => {
+                "Environment has been destroyed. Create a new environment to redeploy.".to_string()
+            }
+            "provision_failed" => {
+                "Provisioning failed. Check error details and retry 'provision'.".to_string()
+            }
+            "configure_failed" => {
+                "Configuration failed. Check error details and retry 'configure'.".to_string()
+            }
+            "release_failed" => {
+                "Release failed. Check error details and retry 'release'.".to_string()
+            }
+            "run_failed" => "Run failed. Check error details and retry 'run'.".to_string(),
+            "destroy_failed" => {
+                "Destruction failed. Check error details and retry 'destroy'.".to_string()
+            }
+            _ => format!("Unknown state: {state_name}. Check environment state file."),
+        }
     }
 }
 
@@ -167,7 +203,7 @@ mod tests {
             "Created".to_string(),
             "LXD".to_string(),
             test_timestamp(),
-            "Next: Run 'provision test-env' to create infrastructure".to_string(),
+            "created".to_string(),
         );
 
         let output = EnvironmentInfoView::render(&info);
@@ -176,7 +212,7 @@ mod tests {
         assert!(output.contains("State: Created"));
         assert!(output.contains("Provider: LXD"));
         assert!(output.contains("Created: 2025-01-07 12:30:45 UTC"));
-        assert!(output.contains("Next: Run 'provision test-env'"));
+        assert!(output.contains("Run 'provision' to create infrastructure."));
     }
 
     #[test]
@@ -186,7 +222,7 @@ mod tests {
             "Provisioned".to_string(),
             "LXD".to_string(),
             test_timestamp(),
-            "Next: Run 'configure prod-env' to install software".to_string(),
+            "provisioned".to_string(),
         )
         .with_infrastructure(InfrastructureInfo::new(
             IpAddr::V4(Ipv4Addr::new(10, 140, 190, 171)),
@@ -213,7 +249,7 @@ mod tests {
             "Running".to_string(),
             "LXD".to_string(),
             test_timestamp(),
-            "Tracker is running!".to_string(),
+            "running".to_string(),
         )
         .with_services(ServiceInfo::new(
             vec!["udp://10.0.0.1:6969/announce".to_string()],
@@ -242,7 +278,7 @@ mod tests {
             "Running".to_string(),
             "LXD".to_string(),
             test_timestamp(),
-            "Tracker is running!".to_string(),
+            "running".to_string(),
         )
         .with_infrastructure(InfrastructureInfo::new(
             IpAddr::V4(Ipv4Addr::new(192, 168, 1, 100)),
@@ -276,7 +312,7 @@ mod tests {
             "Provisioned".to_string(),
             "LXD".to_string(),
             test_timestamp(),
-            "Next step".to_string(),
+            "provisioned".to_string(),
         )
         .with_infrastructure(InfrastructureInfo::new(
             IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1)),
@@ -288,5 +324,34 @@ mod tests {
         let output = EnvironmentInfoView::render(&info);
 
         assert!(output.contains("-p 2222"));
+    }
+
+    mod get_next_step_guidance {
+        use super::*;
+
+        #[test]
+        fn it_should_guide_from_created_state() {
+            let guidance = EnvironmentInfoView::get_next_step_guidance("created");
+            assert!(guidance.contains("provision"));
+        }
+
+        #[test]
+        fn it_should_guide_from_provisioned_state() {
+            let guidance = EnvironmentInfoView::get_next_step_guidance("provisioned");
+            assert!(guidance.contains("configure"));
+        }
+
+        #[test]
+        fn it_should_guide_from_running_state() {
+            let guidance = EnvironmentInfoView::get_next_step_guidance("running");
+            assert!(guidance.contains("test"));
+        }
+
+        #[test]
+        fn it_should_handle_failed_states() {
+            let guidance = EnvironmentInfoView::get_next_step_guidance("provision_failed");
+            assert!(guidance.contains("failed"));
+            assert!(guidance.contains("retry"));
+        }
     }
 }
